requirements.txt
numpy
librosa
soundfile
openai-whisper
sentence-transformers
yake
pyvi
faiss-cpu
mysql-connector-python
yt-dlp
speechbrain
torch
tqdm
Unidecode
